+++ src/main/java/org/geowe/client/local/layermanager/tool/export/VectorFormat.java
@@ -35,6 +35,8 @@ public class VectorFormat {
 			"TopoJSON", "x-topo+json");
 	public static final VectorFormat GPX_FORMAT = new VectorFormat(7, "GPX",
 			"application/gpx+xml");
+	public static final VectorFormat GEOJSON_CSS_FORMAT = new VectorFormat(8, "GeoJSON_CSS",
+			"x-geo+json");
 	
 	private final int id;
 	private final String name;
@@ -73,12 +75,14 @@ public class VectorFormat {
 	public static List<VectorFormat> getSupportedImportFormat() {
 		final List<VectorFormat> vectorFormats = new ArrayList<VectorFormat>();
 		vectorFormats.add(GEO_JSON_FORMAT);
+		vectorFormats.add(GEOJSON_CSS_FORMAT);
 		vectorFormats.add(KML_FORMAT);
 		vectorFormats.add(GML_FORMAT);
 		vectorFormats.add(WKT_FORMAT);
 		vectorFormats.add(TOPO_JSON_FORMAT);
 		vectorFormats.add(GPX_FORMAT);
 		vectorFormats.add(CSV_FORMAT);
+		
 		return vectorFormats;
 	}